requirements.txt
setuptools==65.5.0
python-magic==0.4.27
requests==2.28.2